backend/admissions/throttles.py
from rest_framework.throttling import AnonRateThrottle


class ResultsThrottle(AnonRateThrottle):
    scope = "results"

    def allow_request(self, request, view):
        if getattr(view, "throttle_scope", None) != self.scope:
            return True
        return super().allow_request(request, view)


class PaymentThrottle(AnonRateThrottle):
    scope = "payments"

    def allow_request(self, request, view):
        if getattr(view, "throttle_scope", None) != self.scope:
            return True
        return super().allow_request(request, view)
