backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg[binary]==3.2.13
razorpay==1.4.2
gunicorn==22.0.0
whitenoise==6.7.0
setuptools==80.9.0
python-dotenv==1.0.1
